requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.34,<1.0
httpx>=0.28,<1.0
anthropic>=0.42,<1.0
python-dotenv>=1.0,<2.0
google-api-python-client>=2.100,<3.0
google-auth>=2.20,<3.0
